apps/backend/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
cryptography==44.0.2
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.115.12
flatbuffers==25.2.10
greenlet==3.1.1
h11==0.14.0
html2text==2025.4.15
httpcore==1.0.8
httpx==0.28.1
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
jiter==0.9.0
magika==0.6.1
Markdown==3.8
markdownify==1.1.0
markitdown[all]==0.1.2
mpmath==1.3.0
numpy==2.2.4
onnxruntime==1.21.1
openai==1.75.0
packaging==25.0
pdfminer.six==20250327
protobuf==6.30.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.0
sympy==1.13.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.4.0
uvicorn==0.34.0
alembic==1.14.0
asyncpg==0.30.0
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
